+++ src/intorods/__init__.py
@@ -1,1 +1,7 @@
-__VERSION__ = "0.0.15"
+__VERSION__ = "0.0.16"
+
+import intorods.filesys.fs_ftp
+import intorods.filesys.fs_irods
+import intorods.filesys.fs_local
+import intorods.filesys.fs_scp
+import intorods.filesys.fs_sftp